Break warehouse.py into individual depot files
