+++ tradepy/collectors/market_index.py
@@ -2,7 +2,7 @@ import tradepy
 import pandas as pd
 from tradepy import LOG
 from tradepy.collectors import DataCollector
-from tradepy.warehouse import BroadBasedIndexBarsDepot, SectorIndexBarsDepot
+from tradepy.depot.index import SectorIndexBarsDepot, BroadBasedIndexBarsDepot
 from tradepy.conversion import broad_index_code_name_mapping
