+++ tradepy/collectors/release_restricted_shares.py
@@ -1,6 +1,6 @@
 import tradepy
 from tradepy.collectors import DataCollector
-from tradepy.warehouse import RestrictedSharesReleaseDepot
+from tradepy.depot.misc import RestrictedSharesReleaseDepot
 
 
 class EastMoneyRestrictedSharesReleaseCollector(DataCollector):
