+++ tradepy/depot/etf.py
@@ -1,0 +1,5 @@
+from tradepy.depot.bases import GenericListingDepot
+
+
+class ETFListingDepot(GenericListingDepot):
+    file_name = "etf-listing.csv"
